backend/pyproject.toml
[project]
name = "netcheck-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aio-pika>=9.5.5",
    "alembic>=1.16.2",
    "apscheduler>=3.11.0",
    "asyncpg>=0.30.0",
    "cryptography>=45.0.4",
    "fastapi>=0.115.13",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "rmq-service",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv.sources]
rmq-service = { git = "https://github.com/imbalet/rmq_service" }

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "black>=25.9.0",
]
